refactor and fix error check
