+++ ecommerce/frontend/src/components/messages/Conversation.less
@@ -1,0 +1,3 @@
+.conversation-message-list {
+    height: 100%;
+}
