+++ ecommerce/frontend/src/components/messages/ConversationList.less
@@ -1,0 +1,2 @@
+.conversations-list {
+}
