+++ ecommerce/frontend/src/components/messages/Conversations.less
@@ -6,10 +6,6 @@
     min-height: 500px;
 }
 
-.message-list {
-    height: 100%;
-}
-
 .conversations-list {
 }
