+++ ecommerce/frontend/src/context/ChatContext.js
@@ -14,7 +14,7 @@ function useChat() {
         setIsLoading(true)
         try {
             const { data } = await api.get(`/messages`)
-            if (!data.successful) throw new Error(data)
+            if (!data?.status?.successful) throw new Error(data)
             const conversations = data.conversations.map(formatConversation)
 
             setConversations(conversations)
@@ -28,13 +28,14 @@ function useChat() {
     }
 
     const sendMessage = async message => {
+        console.log('sendMessage message', message)
         setIsLoading(true)
         try {
             const { data } = await api.post(`/messages`, message)
-            if (!data.successful) throw new Error(data)
+            if (!data?.status?.successful) throw new Error(data)
             await getConversations()
         } catch (error) {
-            notification.error({ message: 'There was an error while fetching messages' })
+            notification.error({ message: 'There was an error while sending message' })
             console.error(error)
         } finally {
             setIsLoading(false)
